Added validations, put inputs in same column, deleted class Novel and Anthology
1. Made alert to notify if title or author for the novel is not entered;
2. Made alert to notify if title, editor or atleast 2 stories for the
anthology are not entered;
3. Removed publisher and year of Publication columns and added
Publishing information column containing both;
4. Removed author and editor columns and added them in author/editor
column that shows author for novels and editor for anthology;
5. Removed Series/Series# and Stories columns and added new column
Additional Information containing series(# ) if it's a novel or Number
of stories and their author/s if it's an anthology;
6. Added the displayStoryAuthors() function that returns ${author} or
${numberOfAuthors} authorsdepending if there is only one author or more
then one authors;
7. Created 2 new variables infoNovel and infoAnthology. If the book is
novel additionalInformation gets the value of infoNovel or if anthology
then gets the value of infoAnthology;
8. Removed the Novel and Anthology classes. Made it so there is no need
for more classes then one, the Book class.
9. Removed the Year of Publication input and put select with search. The
options for the new select are generated with a newly added function
optionsGenerator();
10. Added the number of pages validation (from 1 to 1000). Now on input
change there is a function checking if the number is in the range and
giving alert messages accordingly;
11. Added a new div called alertEmptyInputs that is under the
author/editor input and when the title,author or editor are empty, the
alert goes red under them.
12. Added a new div called alertPages that is under the pages input and
when the page number is <1 or >1000 it goes red;
13. Added a new div called alertSuccess that is under the Save Button
and when the book is saved successfully it goes green;
14. Put the jquery-confirm.js and jquery-confirm.css scripts for the
Delete Book button alert comfirmation, put all the code for deleteBook
into a jquery confirm alert;